+++ contracts/agency/libraries/FixedPoint96.sol
@@ -4,5 +4,5 @@ pragma solidity >=0.8.4 <0.9.0;
 library FixedPoint96 {
   // solhint-disable
   uint8 internal constant RESOLUTION = 96;
-  uint256 internal constant Q96 = 0x1000000000000000000000000;
+  uint internal constant Q96 = 0x1000000000000000000000000;
 }
