+++ contracts/agency/peripherals/Keep3rAccountance.sol
@@ -11,26 +11,26 @@ abstract contract Keep3rAccountance is IKeep3rAccountance {
   EnumerableSet.AddressSet internal _keepers;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => uint256) public override workCompleted;
+  mapping(address => uint) public override workCompleted;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => uint256) public override firstSeen;
+  mapping(address => uint) public override firstSeen;
 
   /// @inheritdoc IKeep3rAccountance
   mapping(address => bool) public override disputes;
 
   /// @inheritdoc IKeep3rAccountance
   /// @notice Mapping (job => bonding => amount)
-  mapping(address => mapping(address => uint256)) public override bonds;
+  mapping(address => mapping(address => uint)) public override bonds;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => mapping(address => uint256)) public override jobTokenCredits;
+  mapping(address => mapping(address => uint)) public override jobTokenCredits;
 
   /// @notice The current liquidity credits available for a job
-  mapping(address => uint256) internal _jobLiquidityCredits;
+  mapping(address => uint) internal _jobLiquidityCredits;
 
   /// @notice Map the address of a job to its correspondent periodCredits
-  mapping(address => uint256) internal _jobPeriodCredits;
+  mapping(address => uint) internal _jobPeriodCredits;
 
   /// @notice Enumerable array of Job Tokens for Credits
   mapping(address => EnumerableSet.AddressSet) internal _jobTokens;
@@ -45,16 +45,16 @@ abstract contract Keep3rAccountance is IKeep3rAccountance {
   mapping(address => bool) internal _isKP3RToken0;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => mapping(address => uint256)) public override pendingBonds;
+  mapping(address => mapping(address => uint)) public override pendingBonds;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => mapping(address => uint256)) public override canActivateAfter;
+  mapping(address => mapping(address => uint)) public override canActivateAfter;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => mapping(address => uint256)) public override canWithdrawAfter;
+  mapping(address => mapping(address => uint)) public override canWithdrawAfter;
 
   /// @inheritdoc IKeep3rAccountance
-  mapping(address => mapping(address => uint256)) public override pendingUnbonds;
+  mapping(address => mapping(address => uint)) public override pendingUnbonds;
 
   /// @inheritdoc IKeep3rAccountance
   mapping(address => bool) public override hasBonded;
