+++ contracts/farm/TGRToken.sol
@@ -474,10 +474,11 @@ contract TGRToken is Node, Ownable, ITGRToken, SessionRegistrar, SessionFees, Se
         // Interpretation: TGR tokens not in Cyberswap accounts (tgrftm and tgrhtz), and not in Agency account (votes account), 
         // will be burned at the above rate and interval.
 
-        uint256 decayPer1e12 = _getDecayPer1e12(user_burn); // smaller than its real value.
+        uint decayPer1e12 = _getDecayPer1e12(user_burn); // smaller than its real value.
         if (decayPer1e12 > 0) {
-            uint256 net_collective = _safeSubtract(user_burn.sum_tokens, user_burn.pending_burn);
-            uint256 new_burn = net_collective * decayPer1e12 / 1e12;  // smaller than its real value
+            uint net_collective = _safeSubtract(user_burn.sum_tokens, user_burn.pending_burn);
+            
+            uint new_burn = net_collective * decayPer1e12 / 1e12;  // smaller than its real value
             user_burn.pending_burn += new_burn;
             if (user_burn.pending_burn > user_burn.sum_tokens) {
                 user_burn.pending_burn = user_burn.sum_tokens;
